fix: when federation is disabled, local table APIs dont fail

+++ lib/Controller/RowOCSController.php
@@ -16,6 +16,7 @@ use OCA\Tables\Helper\ConversionHelper;
 use OCA\Tables\Middleware\Attribute\RequirePermission;
 use OCA\Tables\Model\RowDataInput;
 use OCA\Tables\ResponseDefinitions;
+use OCA\Tables\Service\ConfigService;
 use OCA\Tables\Service\FederationService;
 use OCA\Tables\Service\RowService;
 use OCA\Tables\Service\TableService;
@@ -42,6 +43,7 @@ class RowOCSController extends AOCSController {
 		private TableService $tableService,
 		private ViewService $viewService,
 		private FederationService $federationService,
+		private ConfigService $configService,
 	) {
 		parent::__construct($request, $logger, $n, $userId);
 	}
@@ -75,16 +77,8 @@ class RowOCSController extends AOCSController {
 		$tableId = $viewId = null;
 		if ($iNodeType === Application::NODE_TYPE_TABLE) {
 			$tableId = $nodeId;
-			if ($this->federationService->isNodeFederated($tableId, 'table')) {
-				$table = $this->tableService->find($nodeId, true);
-				return new DataResponse($this->federationService->createRow($table, $data));
-			}
 		} elseif ($iNodeType === Application::NODE_TYPE_VIEW) {
 			$viewId = $nodeId;
-			if ($this->federationService->isNodeFederated($viewId, 'view')) {
-				$view = $this->viewService->find($nodeId, false, $this->userId);
-				return new DataResponse($this->federationService->createRow($view, $data));
-			}
 		}
 
 		$newRowData = new RowDataInput();
@@ -93,6 +87,16 @@ class RowOCSController extends AOCSController {
 		}
 
 		try {
+			if ($this->configService->isFederationEnabled()) {
+				if ($tableId !== null && $this->federationService->isNodeFederated($tableId, 'table')) {
+					$table = $this->tableService->find($nodeId, true);
+					return new DataResponse($this->federationService->createRow($table, $data));
+				}
+				if ($viewId !== null && $this->federationService->isNodeFederated($viewId, 'view')) {
+					$view = $this->viewService->find($nodeId, false, $this->userId);
+					return new DataResponse($this->federationService->createRow($view, $data));
+				}
+			}
 			return new DataResponse($this->rowService->create($tableId, $viewId, $newRowData)->jsonSerialize());
 		} catch (BadRequestError $e) {
 			return $this->handleBadRequestError($e);
@@ -134,18 +138,20 @@ class RowOCSController extends AOCSController {
 		$tableId = $viewId = null;
 		if ($iNodeType === Application::NODE_TYPE_TABLE) {
 			$tableId = $nodeId;
-			if ($this->federationService->isNodeFederated($tableId, 'table')) {
-				$table = $this->tableService->find($nodeId, true);
-				return new DataResponse($this->federationService->updateRow($table, $rowId, $data));
-			}
 		} elseif ($iNodeType === Application::NODE_TYPE_VIEW) {
 			$viewId = $nodeId;
-			if ($this->federationService->isNodeFederated($viewId, 'view')) {
-				$view = $this->viewService->find($nodeId, false, $this->userId);
-				return new DataResponse($this->federationService->updateRow($view, $rowId, $data));
-			}
 		}
 		try {
+			if ($this->configService->isFederationEnabled()) {
+				if ($tableId !== null && $this->federationService->isNodeFederated($tableId, 'table')) {
+					$table = $this->tableService->find($nodeId, true);
+					return new DataResponse($this->federationService->updateRow($table, $rowId, $data));
+				}
+				if ($viewId !== null && $this->federationService->isNodeFederated($viewId, 'view')) {
+					$view = $this->viewService->find($nodeId, false, $this->userId);
+					return new DataResponse($this->federationService->updateRow($view, $rowId, $data));
+				}
+			}
 			return new DataResponse($this->rowService->updateSet($rowId, $viewId, $data, $this->userId, $tableId)->jsonSerialize());
 		} catch (NotFoundError $e) {
 			return $this->handleNotFoundError($e);
@@ -177,18 +183,20 @@ class RowOCSController extends AOCSController {
 		$tableId = $viewId = null;
 		if ($iNodeType === Application::NODE_TYPE_TABLE) {
 			$tableId = $nodeId;
-			if ($this->federationService->isNodeFederated($tableId, 'table')) {
-				$table = $this->tableService->find($nodeId, true);
-				return new DataResponse($this->federationService->deleteRow($table, $rowId));
-			}
 		} elseif ($iNodeType === Application::NODE_TYPE_VIEW) {
 			$viewId = $nodeId;
-			if ($this->federationService->isNodeFederated($viewId, 'view')) {
-				$view = $this->viewService->find($nodeId, false, $this->userId);
-				return new DataResponse($this->federationService->deleteRow($view, $rowId));
-			}
 		}
 		try {
+			if ($this->configService->isFederationEnabled()) {
+				if ($tableId !== null && $this->federationService->isNodeFederated($tableId, 'table')) {
+					$table = $this->tableService->find($nodeId, true);
+					return new DataResponse($this->federationService->deleteRow($table, $rowId));
+				}
+				if ($viewId !== null && $this->federationService->isNodeFederated($viewId, 'view')) {
+					$view = $this->viewService->find($nodeId, false, $this->userId);
+					return new DataResponse($this->federationService->deleteRow($view, $rowId));
+				}
+			}
 			return new DataResponse($this->rowService->delete($rowId, $viewId, $this->userId, $tableId)->jsonSerialize());
 		} catch (NotFoundError $e) {
 			return $this->handleNotFoundError($e);
